added 3 Dto files (добавил 3 Dto файла)

+++ student_achievement/src/main/java/kg/mega/student_achievement/models/entities/Scholarship.java
@@ -21,7 +21,7 @@ public class Scholarship {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     Long id;
-    int scholarshipAmount;
+    double scholarshipAmount;
     Date startDate;
     Date endDate;
     boolean active;
